fix bug
make element_byte_size variable to use size_t instead of integer

+++ src/list/list_properties.h
@@ -1,7 +1,9 @@
 #ifndef LIST_PROPERTIES
 #define LIST_PROPERTIES
 
+#include <stddef.h>
+
 extern int size;
-extern int element_byte_size;
+extern size_t element_byte_size;
 
 #endif
